Add repository to composer on Plugin->activate
When the Plugin is activated it adds the package repository that is
needed for ACF PRO to the list of repositories. It adds it at the
beginning to make sure ACF PRO is always installed with this
repository (composer checks repositories in order).
See: https://getcomposer.org/doc/04-schema.md#repositories

The repository definition is saved in a JSON file (instead of being
defined as an array in the Plugin class) to better resemble how it would
look if you just added it to composer.json (more familiar for users).

The version is set to * because it will be replaced by the plugin. The
url is also missing the version and the key because they will be added
by the plugin.