Get version of ACF PRO package from composer.json
The repository that is added by the plugin needs to specify the ACF PRO
package in a version that matches the version constraint of the user. To
ensure that this is given, the plugin uses the version constraint that
the user supplied.

E.g users composer.json:
...
"require": {
  ...
  "advanced-custom-fields/advanced-custom-fields-pro": "1.2.3",
  ...
}
...
--> Version used: 1.2.3

It works if the user specifies the package in require or require-dev. If
the user does not require the package at all, the repository is never
added because it is unnecessary. The case where the user specifies the
package in require AND require-dev is not covered. The plugin will use
the version from the require field (which might not be the expected
behavior for the user). This case was omited because it seems unlikely.