Validate version constraint (major.minor.patch)
The url that is used to download the ACF PRO package can only handle
exact versions with 3 digits (e.g. 1.2.3). That's why the Plugin throws
an error whenever the user uses a different version constraint (that
might be a legal composer version constraint). This error should inform
the user about the problem and make it easy to solve it.

The tests include samples from valid composer version constraints that
can be found here: https://getcomposer.org/doc/articles/versions.md