Get key from environment and add it to dist url
The key should be made available by the user via the environment
variable ACF_PRO_KEY. Since Bedrock works with .env files, the plugin is
also able to use environment variables from .env files (leads to an
easier workflow: key has to be saved into .env once and can be
forgotten).

The installation fails if no key is available in the environment to
provide the user with a helpful error message.