Make plugin work on fresh install and update
The plugin would only work if it was installed before ACF PRO should be
installed. It would fail on a fresh install (because ACF PRO could not
be found in any repository). This means that the user needs to supply a
package repository that contains a desired version, and a dist
url (without version and key). This package has to require the installer
to make sure that the plugin is available before installing ACF PRO.

Another problem was, that the plugin would save the key in
composer.lock. Since it is common practice to put composer.lock under
version control the purpose of the plugin would be defeated. To avoid
this problem, only the version is appended to the url that the users
supplies and the key is only added to the url for the download (with the
PreFileDownload event).

The version however needs to be added to the package. Otherwise the url
for every version would be the same which leads to composer using
already cached files although the might be from a different
version (composer can't tell because the url is exactly the same).

Should work with `composer install`, `composer install` (from
composer.lock), `composer update` and `composer require`.