Improve maxAge for Shields.io badges
The packagist version badge would show an outdated version number. 
The maxAge of 3600 matches the caching time of packagist.

Changed the maxAge of Travis and Coveralls as well because they 
might change as well.

The license badge can stay at the high max Age because it is not going to 
change very often.