Allow 2 digits in the patch part of the version
Version 5.3.10 contains two digits in the patch part of the exact
version and could not be installed due to an UnexpectedValueException
because the version regex only allowed single digits.

I decided to only extend the patch part of the version regex to allow 2
digits because the benefits of catching mistyped version numbers that do
not exist (5.3.10 is the first version with a multiple digit number)
outweights the flexibility of allowing 2 digits in the major and minor
parts as well. This might have to change if 5.10.*.* or 10.*.*.*
approaches.

Fixes #4