Update gist link in README
The original gist link is no longer available.

I originally found this link in the following ACF support forum topic:
https://support.advancedcustomfields.com/forums/topic/composer-support/

This topic also contains a link to an improved fork of this gist. I
added the link of the fork to the README because it also contains the
original code in its revisions.
https://gist.github.com/fThues/705da4c6574a4441b488/revisions

Fixes #7