Doesn't work with plugin Hirak/prestissimo
When using this plugin in conjunction with hirak/prestissimo, the download of acf-pro fails because the custom RemoteFileSystem is overwritten. In order to get it working, we should ensure that this plugins runs *after* hirak/prestissimo.
As both are running with a priority set to 0, setting this one to -1 resolves the issue.